Expirements pipline

+++ src/utils/audio_features/audio_features.py
@@ -70,7 +70,7 @@ class FrequencyDomainFeatures:
     return magnitude, frequency
 
   def stft(self, signal: np.ndarray, frame_length: int, hop_length: int, log_scale: bool = False):
-    s_scale = librosa.stft(signal, n_fft=frame_length, hop_length=hop_length)
+    s_scale = librosa.stft(signal, n_fft=frame_length, hop_length=hop_length, dtype='complex64')
     return np.abs(s_scale) ** 2 if log_scale == True else s_scale
 
   def melfilters(self, sr: int, frame_length: int, n_mels: int = 128):
